+++ src/functions/relu.rs
@@ -30,7 +30,7 @@ impl FunctionImpl for ReLu {
         let x = inputs[0].borrow();
         let mut output = outputs[0].borrow_mut();
 
-        for i in 0..x.size() as usize {
+        for i in 0..x.size() {
             output.data[i] = if x.data[i] > 0.0 { x.data[i] } else { 0.0 };
         }
     }
@@ -45,7 +45,7 @@ impl FunctionImpl for ReLu {
         let mut x = inputs[0].borrow_mut();
         let output = outputs[0].borrow();
 
-        for i in 0..x.size() as usize {
+        for i in 0..x.size() {
             x.grad[i] += if x.data[i] > 0.0 { output.grad[i] } else { 0.0 };
         }
     }
